requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch
pillow